renamed animation files to be more apropriate
