+++ src/app_includes.h
@@ -1,3 +1,4 @@
 #pragma once
 
+#define EDITOR 1
 #define SHOW_LIGHTS 1
